+++ src/sentry/integrations/slack/message_builder/footer.py
@@ -2,14 +2,12 @@ from __future__ import annotations
 
 from dataclasses import dataclass
 from typing import Sequence
-from urllib.parse import urlencode
 
 from sentry import roles
 from sentry.models import Environment, Group, OrganizationMember, Team, User
 from sentry.notifications.notifications.base import BaseNotification
-from sentry.notifications.types import NotificationSettingTypes, get_notification_setting_type_name
+from sentry.notifications.utils.urls import get_settings_url
 from sentry.types.integrations import ExternalProviders
-from sentry.utils.http import absolute_uri
 
 # Defined as a variable so that other providers can more easily copy this file.
 PROVIDER = ExternalProviders.SLACK
@@ -26,6 +24,7 @@ class FooterLink:
 def build_link(link: FooterLink, provider: ExternalProviders) -> str:
     if provider == ExternalProviders.SLACK:
         return f"<{link.url}|{link.label}>"
+    raise RuntimeError("Invalid Provider")
 
 
 def process_footer_part(
@@ -70,37 +69,12 @@ def get_environment_name(group: Group | None) -> str | None:
     except Environment.DoesNotExist:
         return None
 
-    return getattr(environment, "name", None)
-
-
-def get_team_settings_url(team: Team) -> str:
-    return f"/settings/{team.organization.slug}/teams/{team.slug}/notifications/"
-
-
-def get_user_settings_url(type_key: NotificationSettingTypes | None = None) -> str:
-    if not type_key:
-        return "/settings/account/notifications/"
-
-    fine_tuning_key = get_notification_setting_type_name(type_key)
-    return f"/settings/account/notifications/{fine_tuning_key}"
-
-
-def get_settings_url(
-    notification: BaseNotification,
-    recipient: Team | User,
-    provider: ExternalProviders,
-) -> str:
-    if isinstance(recipient, Team):
-        team = Team.objects.get(id=recipient.id)
-        url = get_team_settings_url(team)
-    else:
-        url = get_user_settings_url(notification.notification_setting_type)
-
-    params = notification.get_tracking_params(provider, recipient)
-    return f"{absolute_uri(url)}?{urlencode(params)}"
+    environment_name: str | None = getattr(environment, "name", None)
+    return environment_name
 
 
 def get_role_string(member: OrganizationMember) -> str:
     if member.role == "billing":
         return "Billing Admin"
-    return roles.get(member.role).name
+    role_string: str = roles.get(member.role).name
+    return role_string
